+++ src/PhpSpreadsheet/Chart/DataSeriesValues.php
@@ -446,7 +446,7 @@ class DataSeriesValues extends Properties
                 }
                 unset($dataValue);
             } else {
-                [$worksheet, $cellRange] = Worksheet::extractSheetTitle($this->dataSource, true);
+                [, $cellRange] = Worksheet::extractSheetTitle($this->dataSource, true);
                 $dimensions = Coordinate::rangeDimension(str_replace('$', '', $cellRange ?? ''));
                 if (($dimensions[0] == 1) || ($dimensions[1] == 1)) {
                     $this->dataValues = Functions::flattenArray($newDataValues);