+++ tests/PhpSpreadsheetTests/CustomReader.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhpOffice\PhpSpreadsheetTests;
+
+use PhpOffice\PhpSpreadsheet\Reader\Xlsx as XlsxReader;
+
+// Used in IOFactoryRegister tests
+class CustomReader extends XlsxReader
+{
+}